app/requirements.txt
cryptography==45.0.5
Flask==3.1.1
Flask_Bootstrap==3.3.7.1
flask_cors==6.0.0
flask_limiter==3.12
Flask_SocketIO==5.5.1
flask_wtf==1.2.2
icecream==2.1.4
PyJWT==2.10.1
PyJWT==2.10.1
pymongo==4.13.2
python-dotenv==1.1.1
python_bcrypt==0.3.2
WTForms==3.2.1
